# packages/shared/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/profiles/search": {
      "get": {
        "operationId": "ProfilesController_search",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
...
    "/profiles/{id}": {
      "get": {
        "operationId": "ProfilesController_getById",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      },
      "patch": {
        "operationId": "ProfilesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
...
    "/accounts/register": {
      "post": {
        "operationId": "AccountsController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
    "/accounts/profiles/create": {
      "post": {
        "operationId": "AccountsController_createProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
    "/accounts/profiles/switch/{id}": {
      "patch": {
        "operationId": "AccountsController_switchProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the newly switched destiny profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "CouplApp API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
...
      "ProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "picture": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string"
          },
          "bannerPicture": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "object",
            "nullable": true
          },
          "bio": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "isPrivate": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "picture",
          "bannerPicture",
          "isPrivate"
        ]
      },
      "ProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "picture": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string"
          },
          "bannerPicture": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "object",
            "nullable": true
          },
          "bio": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "isPrivate": {
            "type": "object"
          },
          "isOwner": {
            "type": "boolean"
          },
          "friendship": {
            "type": "object",
            "nullable": true
          },
          "friends": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "picture",
          "bannerPicture",
          "isPrivate",
          "isOwner"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateProfileDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}